+++ sdk/storage/azure-storage-blob/HISTORY.md
@@ -6,7 +6,8 @@
 
 - `set_container_access_policy` has required parameter `signed_identifiers`.
 - NoRetry policy has been removed. Use keyword argument `retry_total=0` for no retries. 
-
+- `AccountSasPermissions`, `BlobSasPermissions`, `ContainerSasPermissions`, `ResourceTypes`, and `Services` 
+now have method `from_string` which takes parameters as a string.
 
 ## Version 12.0.0b4: