+++ sdk/storage/azure-storage-file/HISTORY.md
@@ -7,7 +7,8 @@
 - `file_permission_key` parameter has been renamed to `permission_key`
 - `set_share_access_policy` has required parameter `signed_identifiers`.
 - NoRetry policy has been removed. Use keyword argument `retry_total=0` for no retries. 
-
+- `AccountSasPermissions`, `FileSasPermissions`, `ShareSasPermissions`, `ResourceTypes`, `NTFSAttributes`, 
+and `Services` now have method `from_string` which takes parameters as a string.
 
 ## Version 12.0.0b4: